+++ python/sglang/srt/distributed/device_communicators/pynccl_allocator.py
@@ -405,3 +405,33 @@ def debug_check_symmetric_mempool(
                 "\n".join(bad_details),
                 stack,
             )
+
+
+def prealloc_symmetric_memory_pool(
+    *,
+    is_draft_worker: bool,
+    enable_symm_mem: bool,
+    device: str,
+    forward_stream: torch.cuda.Stream,
+):
+    # PyTorch mempools never de-fragment memory in OOM scenarios, so we need to pre-allocate a large chunk of memory to limit fragmentation.
+    if (
+        is_draft_worker
+        or not enable_symm_mem
+        or envs.SGLANG_SYMM_MEM_PREALLOC_GB_SIZE.get() <= 0
+    ):
+        return
+
+    from sglang.srt.distributed import get_tp_group
+
+    # Memory allocation is tied to a cuda stream, use the forward stream
+    with torch.get_device_module(device).stream(forward_stream):
+        logger.info(
+            f"Pre-allocating symmetric memory pool with {envs.SGLANG_SYMM_MEM_PREALLOC_GB_SIZE.get()} GiB"
+        )
+        with use_symmetric_memory(get_tp_group()):
+            torch.empty(
+                (envs.SGLANG_SYMM_MEM_PREALLOC_GB_SIZE.get() * 1024 * 1024 * 1024,),
+                dtype=torch.uint8,
+                device=device,
+            )
